@@ -1,51 +1,59 @@
 /*
- * Copyright (C) 2018
- *  Source code is created by Elissa Software
- *
+ * Copyright 2019 Wai Yan (TechBase Software). All rights reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and limitations under the License.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 apply plugin: 'com.android.application'
+apply plugin: 'com.jakewharton.butterknife'
 
 repositories {
     maven { url 'https://maven.fabric.io/public' }
 }
 
 apply plugin: 'io.fabric'
+apply plugin: 'com.google.gms.google-services'
 
 buildscript {
     repositories {
         maven { url 'https://maven.fabric.io/public' }
     }
-
     dependencies {
-        classpath 'io.fabric.tools:gradle:1+'
+        classpath 'io.fabric.tools:gradle:1.28.0'
     }
 }
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
-        applicationId "wyp.mmcomputerdictionary"
+        applicationId "wyp.mcd"
         minSdkVersion 19
-        targetSdkVersion 27
-        versionCode 15
-        versionName "2.5"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 18
+        versionName "1.18"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
+
+        javaCompileOptions {
+            annotationProcessorOptions {
+                includeCompileClasspath false
+            }
+        }
     }
 
     signingConfigs {
         release {
-            keyAlias 'mcd-key'
+            keyAlias 'mcd'
             keyPassword 'password'
-            storeFile file('/Users/elissa/Downloads/me/bitbucket/android-projects/mcd/mcd_android/android_keystore.jks')
+            storeFile file('../keystore/mcd-keystore.jks')
             storePassword 'password'
         }
     }
