@@ -1,14 +1,14 @@
 /*
- * Copyright (C) 2018
- *  Source code is created by Elissa Software
- *
+ * Copyright 2019 Wai Yan (TechBase Software). All rights reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and limitations under the License.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package wyp.mcd;
@@ -16,8 +16,9 @@ package wyp.mcd;
 import android.app.Application;
 import android.content.Context;
 import android.os.Build;
-import android.support.multidex.MultiDex;
-import android.support.v7.app.AppCompatDelegate;
+
+import androidx.multidex.MultiDex;
+import androidx.appcompat.app.AppCompatDelegate;
 
 import com.crashlytics.android.Crashlytics;
 import com.crashlytics.android.answers.Answers;
@@ -42,7 +43,7 @@ public class McdApplication extends Application {
     public void onCreate() {
         super.onCreate();
         Fabric.with(this, new Crashlytics());
-        Fabric.with(this, new Answers());
+        Fabric.with(this, new Answers(), new Crashlytics());
         instance = this;
 
         /* SharedPreferences init */
@@ -58,8 +59,9 @@ public class McdApplication extends Application {
         /* Setup notification for app */
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             Logger.d(this.getClass(), "Registering notification channel...");
-            McdNotificationManager.createNotificationChannel(
-                    this, McdNotificationManager.CHAT_CHANNEL_ID, "Notify Incoming",
+            McdNotificationManager.createNotificationChannel(this,
+                    McdNotificationManager.CHAT_CHANNEL_ID,
+                    "Notify Incoming",
                     "Application will notify whenever there is incoming message");
         }
     }
