@@ -1,22 +1,23 @@
 /*
- * Copyright (C) 2018
- *  Source code is created by Elissa Software
- *
+ * Copyright 2019 Wai Yan (TechBase Software). All rights reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and limitations under the License.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package wyp.mcd.infrastructure.pesistance;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
-import android.arch.persistence.room.TypeConverters;
+import androidx.room.Database;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
+import androidx.room.TypeConverters;
+
 import android.content.Context;
 
 import wyp.mcd.component.persistence.DateRoomConverter;
@@ -28,8 +29,12 @@ import wyp.mcd.infrastructure.entities.BookmarksEntity;
 import wyp.mcd.infrastructure.entities.EngToEngEntity;
 import wyp.mcd.infrastructure.entities.EngToMmEntity;
 
-@Database(entities = {BookmarksEntity.class, EngToMmEntity.class, EngToEngEntity.class}
-        , version = 1, exportSchema = false)
+@Database(entities = {
+        BookmarksEntity.class,
+        EngToMmEntity.class,
+        EngToEngEntity.class}
+        , version = 1,
+        exportSchema = false)
 @TypeConverters({DateRoomConverter.class})
 public abstract class McdRoomDatabase extends RoomDatabase {
 
@@ -57,11 +62,12 @@ public abstract class McdRoomDatabase extends RoomDatabase {
                 .build();
     }
 
-    public abstract BookmarksDao bookmarksDao();
 
     public abstract EngToMmDao engToMmDao();
 
     public abstract EngToEngDao engToEngDao();
+
+    public abstract BookmarksDao bookmarksDao();
 
     public void cleanUp() {
         INSTANCE = null;
